Separate Simulator, AdafruitBNO055, and AdafruitBNO085 classes into their own files for import
